# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AlertContacts API",
        "description": "API de sécurité personnelle pour protéger et rassurer ses proches. Cette API permet de gérer les zones de sécurité, zones de danger, invitations, relations entre utilisateurs et notifications en temps réel.",
        "contact": {
            "name": "Support AlertContacts",
            "email": "support@alertcontacts.net"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://alertcontacts.net/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://mobile.alertcontacts.net/api",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
        "/auth/firebase-login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Connexion avec Firebase",
                "description": "Authentifie un utilisateur via Firebase et retourne un token d'accès",
                "operationId": "4e81bc9bcbaeb8da72f0572cd76fe247",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idToken",
                                    "userData"
                                ],
                                "properties": {
                                    "idToken": {
                                        "description": "Token ID Firebase",
                                        "type": "string"
                                    },
                                    "userData": {
                                        "required": [
                                            "uid",
                                            "email"
                                        ],
                                        "properties": {
                                            "uid": {
                                                "description": "UID Firebase de l'utilisateur",
                                                "type": "string"
                                            },
                                            "email": {
                                                "description": "Email de l'utilisateur",
                                                "type": "string",
                                                "format": "email"
                                            },
                                            "name": {
                                                "description": "Nom de l'utilisateur",
                                                "type": "string"
                                            },
                                            "picture": {
                                                "description": "URL de la photo de profil",
                                                "type": "string"
                                            },
                                            "phone_number": {
                                                "description": "Numéro de téléphone",
                                                "type": "string"
                                            },
                                            "email_verified": {
                                                "description": "Email vérifié",
                                                "type": "boolean"
                                            },
                                            "provider": {
                                                "description": "Fournisseur d'authentification",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "Token d'accès",
                                                    "type": "string"
                                                },
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "VALIDATION_ERROR"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Invalid request data"
                                                },
                                                "details": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Déconnexion",
                "description": "Déconnecte l'utilisateur et révoque le token d'accès",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Profil utilisateur",
                "description": "Récupère les informations de l'utilisateur connecté",
                "operationId": "bc2032a379aa97707ccc61511897c120",
                "responses": {
                    "200": {
                        "description": "Informations utilisateur récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "UNAUTHORIZED"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "User not authenticated"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Utilisateur",
                "description": "Modèle représentant un utilisateur de l'application",
                "properties": {
                    "id": {
                        "description": "Identifiant unique de l'utilisateur",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom complet de l'utilisateur",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "firebase_uid": {
                        "description": "UID Firebase de l'utilisateur",
                        "type": "string",
                        "example": "firebase_uid_123",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date de création du compte",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum. Utilisez le token obtenu lors de la connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}